✅Feature-server-side-rendering-in-resources
